refactor chat timeline message actions
